requirements.txt
Django==3.2
django-cors-headers==3.7.0
django-phonenumber-field==5.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
Pillow==8.2.0
